Improve Enum decleration tests
